Removing remaining documentation referrences to reshape
ggpcp stated it was using reshape::melt, but it is actually
using reshape2::melt.

The scale arguments to ggorder and ggstructure had a scale
argument that was no longer used (corresponding argument in
ggpcp removed in 030fadf59a0fd896865b28fd9e76da9698164ec2.
Removed this argument (and its documentation which referenced
reshape).